Renamed the file to follow JavaScript test naming conventions (I had defaulted to what unittest uses in Python).
Was having a lot of trouble getting the output to match, eventually
figured out that by passing the args without quotes, the greeter
script was interpreting the last two words as separate arguments and
was only taking the final argument as a string. Adding quotes around
the arg before passing it ensures anything typed into the command
line is passed as one string. Works!
